libcamera: v4l2_device: Map `BITMASK` type to `Unsigned32`
A bitmask is conceptually an unsigned integer, as implied by the documentation:

  The maximum value is interpreted as a __u32,
  allowing the use of bit 31 in the bitmask.

so map it to a 32-bit unsigned integer. Adjust the rkisp1 pipeline handler
accordingly.

+++ src/libcamera/pipeline/rkisp1/rkisp1.cpp
@@ -1479,9 +1479,7 @@ int PipelineHandlerRkISP1::createCamera(MediaEntity *sensor)
 	if (controls.find(RKISP1_CID_SUPPORTED_PARAMS_BLOCKS) != controls.end()) {
 		auto list = param_->getControls({ { RKISP1_CID_SUPPORTED_PARAMS_BLOCKS } });
 		if (!list.empty())
-			supportedBlocks = static_cast<uint32_t>(
-				list.get(RKISP1_CID_SUPPORTED_PARAMS_BLOCKS)
-					.get<int32_t>());
+			supportedBlocks = list.get(RKISP1_CID_SUPPORTED_PARAMS_BLOCKS).get<uint32_t>();
 	} else {
 		LOG(RkISP1, Error)
 			<< "Failed to query supported params blocks. Falling back to defaults.";
